wip: resolve eslint warning
