+++ src/record/import/usecases/add.ts
@@ -3,7 +3,8 @@ import type { LocalRecord } from "../types/record";
 import type { KintoneRecordForParameter } from "../../../kintone/types";
 import type { RecordSchema } from "../types/schema";
 
-import { fieldProcessor, recordReducer } from "./add/record";
+import { fieldProcessor } from "./add/field";
+import { recordReducer } from "./add/record";
 import { AddRecordsError } from "./add/error";
 import { logger } from "../../../utils/log";
 import { ProgressLogger } from "./add/progress";
